requirements.txt
awebus==1.0a6
cycler==0.10.0
filterpy==1.4.5
geographiclib==1.50
geopy==2.1.0
gps==3.19
kiwisolver==1.3.1
matplotlib==3.4.1
mpu6050-raspberrypi==1.1
numpy==1.20.2
Pillow==8.2.0
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.6.3
six==1.16.0
smbus==1.1.post2
